+++ dotnet/src/HtmlBuilder/Core/Elements/BoldElement.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Text;
 
 namespace PucSp.DesignPatterns.HtmlBuilder.Core
 {
@@ -7,7 +8,7 @@ namespace PucSp.DesignPatterns.HtmlBuilder.Core
     {
         private readonly string _text;
 
-        public string Tag => "strong";
+        public override string Tag => "strong";
 
         public BoldElement(string text)
         {
@@ -17,61 +18,20 @@ namespace PucSp.DesignPatterns.HtmlBuilder.Core
             }
 
             this._text = text;
-            this.attributes = new HashSet<HtmlAttribute>();
         }
 
-        public IHtmlElement AddNode(IHtmlElement node)
+        public override string ToHtml()
         {
-            throw new NotImplementedException();
-        }
+            var result = new StringBuilder();
 
-        public IHtmlElement AddAttribute(string name, string value)
-        {
-            if (name == null)
+            foreach (var item in base.nodes)
             {
-                throw new ArgumentNullException(nameof(name));
+                result.Append(item.Tag);
             }
 
-            if (name.Length == 0)
-            {
-                throw new ArgumentException("Attribute name can not be empty.", nameof(name));
-            }
-
-            if (value == null)
-            {
-                throw new ArgumentNullException(nameof(value));
-            }
-
-            return this.AddAttribute(new HtmlAttribute(name, value));
+            return result.ToString();
         }
 
-        public IHtmlElement AddAttribute(HtmlAttribute attribute)
-        {
-            if (attribute == null)
-            {
-                throw new ArgumentNullException(nameof(attribute));
-            }
-
-            base.attributes.Add(attribute);
-
-            return this;
-        }
-
-        public IHtmlElement RemoveAttribute(string name)
-        {
-            base.attributes.RemoveWhere(x => x.Name == name);
-
-            return this;
-        }
-
-        public string ToHtml()
-        {
-            throw new NotImplementedException();
-        }
-
-        protected override string AsTagString()
-        {
-            throw new NotImplementedException();
-        }
+        protected override string AsTagString() => $"<{this.Tag}>{this._text}</{this.Tag}>";
     }
 }